address comment and add docs
