+++ test_app/custom_queries.py
@@ -1,5 +1,5 @@
 import strawberry
-from api.queries import BaseQuery
+from api.queries import Query as BaseQuery
 
 
 @strawberry.field
@@ -9,5 +9,4 @@ def my_custom_field(self) -> str:
 
 @strawberry.type
 class Query(BaseQuery):
-
     foo: str = my_custom_field
